requirements.txt
deep_translator==1.11.4
fitz==0.0.1.dev2
python_pptx==0.6.23
